Add styling of upload list on Welcome screen
Add ability to change styles for upload list component.

MOB-1839
