+++ GliaWidgets/Resources/Assets.xcassets/MessageCenter/Upload/mc-remove-upload.imageset/mc-remove-upload.pdf
@@ -1,0 +1,80 @@
+%PDF-1.7
+
+1 0 obj
+  << /ExtGState << /E1 << /ca 0.300000 >> >> >>
+endobj
+
+2 0 obj
+  << /Length 3 0 R >>
+stream
+/DeviceRGB CS
+/DeviceRGB cs
+q
+/E1 gs
+1.000000 0.000000 -0.000000 1.000000 0.000000 0.000000 cm
+0.380392 0.415686 0.458824 scn
+12.000000 10.791429 m
+10.791429 12.000000 l
+6.000000 7.208571 l
+1.208571 12.000000 l
+0.000000 10.791429 l
+4.791429 6.000000 l
+0.000000 1.208571 l
+1.208571 0.000000 l
+6.000000 4.791429 l
+10.791429 0.000000 l
+12.000000 1.208571 l
+7.208571 6.000000 l
+12.000000 10.791429 l
+h
+f
+n
+Q
+
+endstream
+endobj
+
+3 0 obj
+  404
+endobj
+
+4 0 obj
+  << /Annots []
+     /Type /Page
+     /MediaBox [ 0.000000 0.000000 12.000000 12.000000 ]
+     /Resources 1 0 R
+     /Contents 2 0 R
+     /Parent 5 0 R
+  >>
+endobj
+
+5 0 obj
+  << /Kids [ 4 0 R ]
+     /Count 1
+     /Type /Pages
+  >>
+endobj
+
+6 0 obj
+  << /Pages 5 0 R
+     /Type /Catalog
+  >>
+endobj
+
+xref
+0 7
+0000000000 65535 f
+0000000010 00000 n
+0000000074 00000 n
+0000000534 00000 n
+0000000556 00000 n
+0000000729 00000 n
+0000000803 00000 n
+trailer
+<< /ID [ (some) (id) ]
+   /Root 6 0 R
+   /Size 7
+>>
+startxref
+862
+%%EOF
