+++ GliaWidgetsTests/SecureConversations/SecureConversations.FileUploadListViewModel.Environment.Failing.swift
@@ -1,0 +1,9 @@
+@testable import GliaWidgets
+
+extension SecureConversations.FileUploadListViewModel.Environment {
+    static let failing = SecureConversations.FileUploadListViewModel.Environment(
+        uploader: .failing,
+        style: .chat(.mock),
+        uiApplication: .failing
+    )
+}
